% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBuilding.R
\name{buildBuilding}
\alias{buildBuilding}
\title{Build a building}
\usage{
buildBuilding(
  length = 8,
  width = 6,
  height = 5,
  foundation = 1,
  wall = 45,
  floorBoards = 125,
  carpet = 171,
  carpet_style = 11,
  gap = 0,
  player_id = NULL
)
}
\arguments{
\item{length}{Length of the building}

\item{width}{Width of the building}

\item{height}{Height of the building}

\item{foundation}{ID for foundation blocks (1 = stone)}

\item{wall}{ID for wall blocks (45 = brick)}

\item{floorBoards}{ID for floor boards (125 = wood planks)}

\item{carpet}{ID for carpet blocks (171 = carpet)}

\item{carpet_style}{ID for carpet style (11 = blue)}

\item{gap}{ID for gap blocks (0 = air)}

\item{player_id}{Player ID}
}
\value{
None.
}
\description{
Build a building with specified dimensions, with a foundation, walls, floorboards, carpet, and a door
}
\author{
Felix Ling
}
