% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDoor.R
\name{buildDoor}
\alias{buildDoor}
\title{Builds a door}
\usage{
buildDoor(
  x = NULL,
  y = NULL,
  z = NULL,
  direction = NULL,
  doorBlock = 64,
  pressurePlate = TRUE,
  pressurePlateBlock = 70,
  player_id = NULL
)
}
\arguments{
\item{x}{x-coordinate of where to put the door. If any of \code{x}, \code{y},
or \code{z} are NULL, will calculate from player's position and
direction they are facing.}

\item{y}{y-coordinate of where to put the door.}

\item{z}{z-coordinate of where to put the door.}

\item{direction}{The compass direction the door should open in.}

\item{doorBlock}{The material to create the door out of.}

\item{pressurePlate}{Whether to put pressure plates around the door.}

\item{pressurePlateBlock}{The material to create the pressure plate out of.}

\item{player_id}{Player ID}
}
\description{
Build a door in the wall (or any non-AIR block) that the player is looking
at, opening in that same direction. If desired, also puts a pressure plate in
front of and behind the door to automatically open it. Alternatively, can
specify the coordinates and direction manually.
}
\details{
For door blocks, see \code{mc::find_item("door")}; they include
64 for oak and 71 for iron. Pressure place blocks include 70
for stone and 72 for wood.
}
\author{
Felix Ling
}
