% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRlogo.R
\name{buildRlogo}
\alias{buildRlogo}
\title{Render R logo in minecraft}
\usage{
buildRlogo(lowerleft, height = 80, width = NULL)
}
\arguments{
\item{lowerleft}{Lower left position (}

\item{height}{Height of R logo in blocks}

\item{width}{Width of R logo in blocks
(if not provided, determined to preserve aspect ratio}
}
\description{
Render the R logo in minecraft
}
\examples{
\dontrun{
library(miner)
mc_connect()
pos <- getPlayerPos(getPlayerIds()[1])
buildRlogo(pos + c(5, 10, 5))
}
}
