% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStairs.R
\name{buildStairs}
\alias{buildStairs}
\title{Builds stairs}
\usage{
buildStairs(
  player_id = NULL,
  down = TRUE,
  stopHeight = 0,
  stopOnNotAir = TRUE,
  width = 1,
  stairId = 109,
  blockId = 98,
  blockStyle = 0,
  buildBlock = TRUE
)
}
\arguments{
\item{player_id}{Player ID}

\item{down}{Whether to build stairs downward. If false, builds upward}

\item{stopHeight}{The height at which to stop building.}

\item{stopOnNotAir}{Whether to stop if about to build where something exists.}

\item{width}{How wide to build the staircase. MUST BE AN ODD NUMBER}

\item{stairId}{The block ID for the stairs.}

\item{blockId}{The block ID for the block underneath each stair. Can set to AIR
if you don't want this block.}

\item{blockStyle}{The block style for the block underneath each stair.}

\item{buildBlock}{If false, don't put the blocks under the stairs}
}
\description{
Build stairs up or down from player position in the specified direction.
For now, width should be odd so that we can have a central staircase and
extend to the right and left of it to widen it.
}
\details{
For stairs block IDs, see \code{miner::find_item("stairs")}; they include
53 for oak, 67 for cobblestone, and 108 for brick. The corresponding blocks for these materials are 5 for oak, 4 for cobblestone, and 45 for brick
}
