% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft-package.R
\docType{package}
\name{craft-package}
\alias{craft}
\alias{craft-package}
\title{craft: Extra Functions to Control Minecraft with RaspberryJuice API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extra functions to illustrate the use of the miner package to control Minecraft with the RaspberryJuice API.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/craft/}{Extra functions to control Minecraft}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/craft}
  \item Report bugs at \url{https://github.com/kbroman/craft/issues}
}

}
\author{
\strong{Maintainer}: Karl Broman \email{broman@wisc.edu}

Authors:
\itemize{
  \item Karl Broman \email{broman@wisc.edu}
  \item Brooke Anderson
  \item Gergely Daróczi
  \item Ali Zaidi
  \item Mario Inchiosa
  \item David Smith
  \item Felix Ling
}

}
\keyword{internal}
