% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Make a cube}
\usage{
cube(xlen = 10, ylen = 10, zlen = 10, blockid = 1, styleid = 0,
  fill = FALSE, offset = c(0, -1, 0), playerid = getPlayerIds(), pos,
  xlim, ylim, zlim)
}
\arguments{
\item{xlen}{integer. Cube length.}

\item{ylen}{integer. Cube hight.}

\item{zlen}{integer. Cube depth.}

\item{blockid}{integer. Block type.}

\item{styleid}{integer. Block style.}

\item{fill}{logical. If TRUE then solid. If FALSE then hollow.}

\item{offset}{vector of length three. How many units to offset the cube from \code{pos}.}

\item{playerid}{integer of length one. Defaults to player id in solo play. Set explicitly in multi play.}

\item{pos}{vector of length three. Defines the corner of the cube. Defaults to the player's position.}

\item{xlim}{vector of length two. Defines the lower and upper cuttoff points to truncate the cube.}

\item{ylim}{vector of length two. Defines the lower and upper cuttoff points to truncate the cube.}

\item{zlim}{vector of length two. Defines the lower and upper cuttoff points to truncate the cube.}
}
\value{
Builds a cube or cuboid with the corner at the player's position. By default, the cube is hollow, but you can also use \code{fill=TRUE} to create a solid cube. The players position is determined by \code{getPlayerPos}. You can you can reposition the cube with the \code{offset} command. Use \code{xlim}, \code{ylim}, \code{zlim} to truncate the cube to create a fence or wall. This function will return the cube origin.
}
\description{
Make a solid or hollow cube.
}
\examples{
\dontrun{

# Stone cube 10x10x10
cube(pos = c(0, 0, 0))

# Erase cube
cube(blockid = 0, pos = c(0, 0, 0))

# Stone fence
cube(ylim = c(2, 3), offset = c(0, -2, 0))
}

}
\seealso{
\code{\link{sphere}} to create a sphere.
}
