% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsafy.R
\name{de_elsafy}
\alias{de_elsafy}
\title{Give a player reverse Elsa powers}
\usage{
de_elsafy(player_id = NULL, water = 8, ice = 174, delay = 0.05)
}
\arguments{
\item{player_id}{Player's entity ID}

\item{water}{Item IDs for water}

\item{ice}{Item ID for ice}

\item{delay}{Delay (in seconds) between calls to the minecraft server}
}
\value{
None.
}
\description{
Give a player revsese Elsa powers, to clean up after running \code{\link[=elsafy]{elsafy()}}: when she walks over ice, it turns back to water.
}
\examples{
\dontrun{
# Runs in an infinite loop; press ctrl-C to stop.
de_elsafy(215)
}

}
\seealso{
\code{\link[=elsafy]{elsafy()}}
}
