% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawLine.R
\name{drawLine}
\alias{drawLine}
\title{Draws a line}
\usage{
drawLine(p0, p1, id = 1, style = 0)
}
\arguments{
\item{p0}{vector of the first endpoint}

\item{p1}{vector of the second endpoint}

\item{id}{Minecraft block ID to draw the line with}

\item{style}{Minecraft block style ID to draw the line with (e.g., color)}
}
\value{
None
}
\description{
Implements Bresenham's line-drawing algorithm in 3D space. Takes two points,
where each point is a vector of 3 coordinates, and constructs a line between
the two points.
}
\author{
Felix Ling, based on python code in a gist by \code{theJollySin} that has since been removed
}
