% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsafy.R
\name{elsafy}
\alias{elsafy}
\title{Give a player Elsa powers}
\usage{
elsafy(player_id = NULL, water = c(8, 9), ice = 174, delay = 0.02)
}
\arguments{
\item{player_id}{Player's entity ID}

\item{water}{Vector of item IDs for different kinds of water}

\item{ice}{Item ID for ice}

\item{delay}{Delay (in seconds) between calls to the minecraft server}
}
\value{
None.
}
\description{
Give a player Elsa powers: when she walks on water, it turns to ice.
}
\examples{
\dontrun{
# Runs in an infinite loop; press ctrl-C to stop.
elsafy(215)
}

}
\seealso{
\code{\link[=de_elsafy]{de_elsafy()}}
}
