% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_items.R
\name{find_items}
\alias{find_items}
\title{Find multiple items by name or ID/style}
\usage{
find_items(name = NULL, id = NULL, style = 0)
}
\arguments{
\item{name}{Vector of character string with the names of Minecraft
items (specify either \code{name} or \code{id}, not both)}

\item{id}{Vector of numeric or character strings with the ID of a
Minecraft item (specify either \code{name} or \code{id}, not both)}

\item{style}{Vector of numeric or character string with the style
of a Minecraft item (use this argument only if querying by \code{id}
is provided); should have length 1 or the same length as \code{id}.}
}
\value{
Data frame with a row for each item found in \code{\link[miner:mc_items]{mc_items}},
provided the matches are unique.
}
\description{
Find a set of Minecraft items by name or ID. If querying by ID, the
search can also specify item styles.
}
\details{
If \code{name} is provided, we first look to see whether there
is an exact match to the \code{name} column in \code{\link[miner:mc_items]{mc_items}}. If there
is, we return that row. If not, we don't include the results.

If instead \code{id} is provided, we return the row with that id
and \code{style==0} (or whatever style was provided).
}
\examples{
flower_names <- c("Chorus Flower", "Peony", "Rose Bush",
    "Lilac", "Sunflower", "Pink Tulip", "White Tulip",
    "Orange Tulip", "Red Tulip", "Oxeye Daisy", "Allium",
    "Dandelion", "Poppy", "Blue Orchid", "Peony")
flowers <- find_items(flower_names)

}
\seealso{
\code{\link[miner:find_item]{miner::find_item()}}, \link[miner:mc_items]{miner::mc_items}
}
