% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_sets.R
\docType{data}
\name{font_sets}
\alias{font_sets}
\title{Font set data}
\format{An object of class \code{list} of length 12.}
\source{
\url{http://uzebox.org/wiki/index.php?title=Font_Bitmaps}
}
\usage{
data(font_sets)
}
\description{
Font set data used to construct text from blocks within minecraft
}
\examples{
data(font_sets)
}
\keyword{datasets}
