% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{getBlocksV}
\alias{getBlocksV}
\title{getBlocks but taking vector positions}
\usage{
getBlocksV(pos0, pos1)
}
\arguments{
\item{pos0}{Vector (x,y,z) of first position}

\item{pos1}{Vector (x,y,z) of second position}
}
\value{
An 3-D array of integers where each integer gives the ID of the
type of a block in the cuboid.
}
\description{
Determine blocks within a cuboid
}
\examples{
\dontrun{
library(miner)
mc_connect()
p <- getPlayerPos()
getBlocksV(p - c(0, 1, 0), p - c(0, 5, 0))
}

}
\seealso{
\code{\link[miner:getBlocks]{miner::getBlocks()}}
}
