% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{getBlockV}
\alias{getBlockV}
\title{getBlock but taking a vector position}
\usage{
getBlockV(pos, include_style = TRUE)
}
\arguments{
\item{pos}{Vector (x,y,z) of position}

\item{include_style}{A logical value of whether the block's
style should also be included in the output (defaults to TRUE).}
}
\value{
A numeric vector of length one or two with the type ID
and style, if \code{include_style} is \code{TRUE}, of the block
at position (x, y, z). You can use \code{\link[miner:find_item]{find_item()}} to
find the name of the block type based on this returned ID.
}
\description{
Determine block at position (x,y,z)
}
\examples{
\dontrun{
library(miner)
mc_connect()
p <- getPlayerPos()
getBlockV(p + c(0, -1, 0))
}

}
\seealso{
\code{\link[miner:getBlock]{miner::getBlock()}}
}
