% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerCompass.R
\name{getPlayerCompass}
\alias{getPlayerCompass}
\title{Get player's rotation  as compass bearing}
\usage{
getPlayerCompass(player_id = NULL, n_compass_points = c("16", "8", "4"))
}
\arguments{
\item{player_id}{Numeric ID for the player}

\item{n_compass_points}{Number of compass points to use (4, 8, or 16)}
}
\value{
A character string representing the player's direct as a compass bearing
}
\description{
Get the direction that a player is facing as a compass bearing (north, south, east, west, etc.)
}
\examples{
\dontrun{
library(miner)
getPlayerIds()
getPlayerRotation(355)
getPlayerCompass(355, 8)
}

}
