% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_towers.R
\name{ice_towers}
\alias{ice_towers}
\title{Make a random ice tower at hit locations}
\usage{
ice_towers(player_id, block_id = 212, max_height = 4, delay = 0.2)
}
\arguments{
\item{player_id}{Player's entity ID}

\item{block_id}{Item ID (id=212 for frosted ice)}

\item{max_height}{Maximum height of ice tower}

\item{delay}{Delay (in seconds) between calls to the minecraft server}
}
\value{
None.
}
\description{
Make an ice tower of a random height wherever you hit.
}
\note{
Only right clicks with an iron sword will work.
}
\examples{
\dontrun{
# connect to minecraft
miner::mc_connect()

# Need to use ctrl-c to stop
ice_towers( getPlayerIds()[1] )
}
}
\seealso{
\code{\link[miner:getBlockHits]{miner::getBlockHits()}}
}
