% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_race.R
\name{mc_race}
\alias{mc_race}
\title{Determine who is fastest}
\usage{
mc_race(time = 10)
}
\arguments{
\item{time}{Length of race (in seconds)}
}
\value{
Return the player_id of whoever moves the farthest
}
\description{
Hold a race, started in the chat, and see who moves the farthest
}
\examples{
\dontrun{
library(miner)
mc_connect()
whoisfastest()
}

}
