% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_Reval.R
\name{mc_Reval}
\alias{mc_Reval}
\title{Evaluate R within Minecraft}
\usage{
mc_Reval(player_id = NULL, delay = 1)
}
\arguments{
\item{player_id}{Player's entity ID}

\item{delay}{Delay (in seconds) between checks of the chat messages}
}
\value{
None.
}
\description{
Look for chat messages that start with \code{"R "} and then evaluate the rest as an R expression.
}
\details{
The function looks at chat messages in Minecraft. For messages
that begin \code{"R "}, the rest of the message is treated as an R expression
and is evaluated.

The default is to evaluate any such R code. If \code{player_id} is
specified, only messages from that player are considered.

Use \strong{extreme} caution with this; if you don't specify
\code{player_id} correctly, you could be letting others execute
code on your computer and so giving them the opportunity to do
bad things.
}
\examples{
\dontrun{
# runs in an infinite loop; use ctrl-c to stop
mc_Reval(212)}
}
