% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_whoami.R
\name{mc_whoami}
\alias{mc_whoami}
\title{Listen for "Who am I"}
\usage{
mc_whoami(delay = 1, max_time = Inf)
}
\arguments{
\item{delay}{Delay (in seconds) between checks of the chat messages}

\item{max_time}{Maximum amount of time (in seconds) before quitting.}
}
\value{
A vector of player IDs
}
\description{
Listen for "Who am I" on minecraft chat, and respond with the player's entity ID.
}
\details{
Uses \code{\link[base:grepl]{base::grepl()}}, ignoring case, to look for
"who am I" in the chat messages, and prints the player ID of
any player who posts a chat message containing that phrase.
}
\examples{
\dontrun{
#' will run for 15 seconds
mc_whoami(max_time=15)}
}
