% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_guess.R
\name{num_guess}
\alias{num_guess}
\title{Guess a number game}
\usage{
num_guess(max_num = 100, delay = 0.2)
}
\arguments{
\item{max_num}{Maximum number}

\item{delay}{Delay (in seconds) between calls to the minecraft server}
}
\value{
The correct number
}
\description{
Play guess-a-number game in the chat
}
\examples{
\dontrun{
# connect to minecraft
miner::mc_connect()

# Need to use ctrl-c to stop early
num_guess()
}
}
\seealso{
\code{\link[miner:chatPost]{miner::chatPost()}}, \code{\link[miner:getChatPosts]{miner::getChatPosts()}}b
}
