% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity.R
\name{purity}
\alias{purity}
\title{purity}
\usage{
purity(z, y, k)
}
\arguments{
\item{z}{cluster assignments}

\item{y}{true cluster labels}

\item{k}{number of non-empty clusters}
}
\description{
Computes the purity where cluster assignments are given by z, true labels by y, and number of (non-empty) clusters by k
}
