% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlogo.R
\docType{data}
\name{Rlogo}
\alias{Rlogo}
\title{R logo data}
\format{
An object of class \code{cimg} (inherits from \code{imager_array}, \code{numeric}) of dimension 800 x 700 x 1 x 1.
}
\source{
\url{https://www.r-project.org/logo/"}
}
\usage{
data(Rlogo)
}
\description{
Matrix corresponding to a version of the R logo
}
\details{
The dataset is a matrix of pixel colors for a reduced-size version of the R logo,
with 1=white, 2=gray, and 3=blue.
}
\examples{
data(Rlogo)
image(Rlogo[,ncol(Rlogo):1], col=c("white", "gray", "blue"),
      bty="n", xaxt="n", yaxt="n")
}
\keyword{datasets}
