% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{setBlocksMixV}
\alias{setBlocksMixV}
\title{setBlocksMix but taking vector positions}
\usage{
setBlocksMixV(pos0, pos1, ids, styles = NULL, prob = NULL)
}
\arguments{
\item{pos0}{Vector (x,y,z) of first position}

\item{pos1}{Vector (x,y,z) of second position}

\item{ids}{Vector of block IDs}

\item{styles}{Vector of block styles}

\item{prob}{Probabilities for each block type}
}
\value{
None.
}
\description{
Place a cuboid of blocks of a random mixture of types
}
\examples{
\dontrun{
library(miner)
mc_connect()
p <- getPlayerPos()
setBlocksStyleV(p + c(0, 1, 0), p + c(0, 5, 0), 95, 2)
}

}
\seealso{
\code{\link[=setBlocksMix]{setBlocksMix()}}
}
