% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{setBlocksStyleV}
\alias{setBlocksStyleV}
\title{setBlocksStyle but taking vector positions}
\usage{
setBlocksStyleV(pos0, pos1, id, style)
}
\arguments{
\item{pos0}{Vector (x,y,z) of first position}

\item{pos1}{Vector (x,y,z) of second position}

\item{id}{Block ID}

\item{style}{Block style}
}
\value{
None.
}
\description{
Place a cuboid of blocks of a single type, allowing a style parameter
}
\examples{
\dontrun{
library(miner)
mc_connect()
p <- getPlayerPos()
setBlocksStyleV(p + c(0, 1, 0), p + c(0, 5, 0), 95, 2)
}

}
\seealso{
\code{\link[=setBlocksStyle]{setBlocksStyle()}}
}
