% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{setBlocksV}
\alias{setBlocksV}
\title{setBlocks but taking vector positions}
\usage{
setBlocksV(pos0, pos1, id)
}
\arguments{
\item{pos0}{Vector (x,y,z) of first position}

\item{pos1}{Vector (x,y,z) of second position}

\item{id}{Block ID}
}
\value{
None.
}
\description{
Place a cuboid of blocks of a single type
}
\examples{
\dontrun{
library(miner)
mc_connect()
p <- getPlayerPos()
setBlocksV(p + c(0, 1, 0), p + c(0, 5, 0), 46)
}

}
\seealso{
\code{\link[miner:setBlocks]{miner::setBlocks()}}
}
