% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{setBlockV}
\alias{setBlockV}
\title{setBlock but taking a vector position}
\usage{
setBlockV(pos, id, style = 0)
}
\arguments{
\item{pos}{Vector (x,y,z) of position}

\item{id}{Block ID}

\item{style}{Block style}
}
\value{
None.
}
\description{
Place a block at position (x,y,z) by type id
}
\examples{
\dontrun{
library(miner)
mc_connect()
p <- getPlayerPos()
setBlockV(p + c(0, 5, 0), 46)
}

}
\seealso{
\code{\link[miner:setBlock]{miner::setBlock()}}
}
