% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{setPlayerPosV}
\alias{setPlayerPosV}
\title{setPlayerPos but taking a vector of positions}
\usage{
setPlayerPosV(pos, player_id = NULL, tile = FALSE)
}
\arguments{
\item{pos}{Vector (x,y,z) giving position}

\item{player_id}{Player ID}

\item{tile}{If TRUE, truncation position to integers}
}
\value{
None.
}
\description{
Move player to position (x,y,z) specified as a vector
}
\examples{
\dontrun{
library(miner)
mc_connect()
pos <- getPlayerPos()
setPlayerPosV(pos + c(0, 5, 0))
}

}
\seealso{
\code{\link[miner:setPlayerPos]{miner::setPlayerPos()}}
}
