% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere}
\alias{sphere}
\title{Make a sphere}
\usage{
sphere(radius = 15, blockid = 20, styleid = 0, fill = FALSE,
  offset = c(0, 0, 0), playerid = miner::getPlayerIds(), pos, xlim,
  ylim, zlim)
}
\arguments{
\item{radius}{integer. Specifies the radius of the sphere.}

\item{blockid}{integer. Block type to be used.}

\item{styleid}{integer. Block style to be used.}

\item{fill}{logical. Should the sphere be solid or hollow. Defaults to hollow.}

\item{offset}{vector of length three. Defines where to place the sphere with respect to \code{pos}.}

\item{playerid}{integer of length one. Defaults to player id in solo play. Set explicitly in multi play.}

\item{pos}{vector of length three. Defines the center of the sphere. Defaults to the player id position.}

\item{xlim}{vector of length two. Defines the lower and upper cuttoff points to truncate the sphere.}

\item{ylim}{vector of length two. Defines the lower and upper cuttoff points to truncate the sphere.}

\item{zlim}{vector of length two. Defines the lower and upper cuttoff points to truncate the sphere.}
}
\value{
Builds a sphere (or a truncated sphere) around the players current position. By default, the sphere is hollow, but you can also use \code{fill=TRUE} to create a solid sphere. The players position is determined by \code{\link[miner]{getPlayerPos}}. You can you can reposition the sphere with the \code{offset} command. Use \code{xlim}, \code{ylim}, \code{zlim} to truncate the sphere. This function will return the sphere's origin.
}
\description{
Make a solid or hollow sphere with the player at the center.
}
\examples{
\dontrun{

# Hollow glass sphere of size 15
sphere(pos = c(0, 0, 0))

# Erase sphere
sphere(blockid = 0, pos = c(0, 0, 0))

# Glass dome
sphere(ylim = c(0, 15))
}

}
\seealso{
\code{\link{cube}} to create a cube.
}
