% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_text.R
\name{write_text}
\alias{write_text}
\title{Write text in minecraft}
\usage{
write_text(text, bottomleft, font = c("4x5", "4x6", "4x8", "6x6", "8x6",
  "4x12", "6x12", "8x12", "8x16", "16x16"), id = 1, style = 0,
  dir = c("north", "south", "east", "west", "up", "down"), top = c("up",
  "east", "west", "north", "south", "down"))
}
\arguments{
\item{text}{A character string to be written in blocks}

\item{bottomleft}{Vector of length 3 representing the location for the lower-left corner for the text.}

\item{font}{Font size to use}

\item{id}{Block type to use to write the text}

\item{style}{Block style to use to write the text}

\item{dir}{Direction the text should run}

\item{top}{Direction that the top of the text should point}
}
\value{
None
}
\description{
Write some text in minecraft, using blocks
}
\examples{
\dontrun{
library(miner)
v <- getPlayerPos()
write_text("Hello!", v+c(5,3,0), font="4x8")
}

}
