% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFence.R
\name{buildFence}
\alias{buildFence}
\title{Build a fence around a square area}
\usage{
buildFence(
  length = 8,
  fenceBlock = 85,
  gateBlock = 107,
  foundationBlock = 1,
  player_id = NULL
)
}
\arguments{
\item{length}{Length and width of fence (must be >= 3)}

\item{fenceBlock}{Block ID for fence}

\item{gateBlock}{Block ID for gate in fence}

\item{foundationBlock}{Block ID for foundation (to get everything to a uniform height)}

\item{player_id}{Player ID; fence centered at player's current position}
}
\value{
None.
}
\description{
Build a fence around a square area, adding a gate in a random side, and
filling the ground underneath to make a uniform height
}
