% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFence.R
\name{buildFence}
\alias{buildFence}
\title{Build a fence around a square area}
\usage{
buildFence(length = 8, fenceBlock = 85, gateBlock = 107, player_id = NULL)
}
\arguments{
\item{length}{Length and width of fence}

\item{fenceBlock}{Block ID for fence}

\item{gateBlock}{Block ID for gate in fence}

\item{player_id}{Player ID; fence centered at player's current position}
}
\value{
None.
}
\description{
Build a fence around a square area, adding a gate at a random side, and
filling the ground to make a uniform height
}
