% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRlogo.R
\name{buildRlogo}
\alias{buildRlogo}
\title{Render R logo in minecraft}
\usage{
buildRlogo(
  bottomleft,
  height = 80,
  width = NULL,
  dir = c("north", "south", "east", "west")
)
}
\arguments{
\item{bottomleft}{Bottom left position (}

\item{height}{Height of R logo in blocks}

\item{width}{Width of R logo in blocks
(if not provided, determined to preserve aspect ratio}

\item{dir}{Which direction should the logo go?}
}
\description{
Render the R logo in minecraft
}
\examples{
\dontrun{
library(miner)
mc_connect()
pos <- getPlayerPos(getPlayerIds()[1])
buildRlogo(pos + c(5, 10, 5))
}
}
