% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearSpace.R
\name{clearSpace}
\alias{clearSpace}
\title{Clear out a cube around a player}
\usage{
clearSpace(length, player_id = NULL)
}
\arguments{
\item{length}{length of each side of the cube to be cleared}

\item{player_id}{player's entity id}
}
\description{
Clear out a cube around a player
}
\examples{
\dontrun{
id <- getPlayerIds()[1]
clearSpace(5, id)
}
}
\seealso{
\code{\link[miner:setBlocks]{miner::setBlocks()}}, \code{\link[miner:getPlayerPos]{miner::getPlayerPos()}}
}
\author{
Felix Ling
}
