% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_sets.R
\docType{data}
\name{font_sets}
\alias{font_sets}
\title{Font set data}
\format{
An object of class \code{list} of length 12.
}
\source{
\url{http://uzebox.org/wiki/index.php?title=Font_Bitmaps}
}
\usage{
data(font_sets)
}
\description{
Font set data used to construct text from blocks within minecraft
}
\details{
The dataset is a list of fonts with each font being represented as a list with
\itemize{
\item \code{charset} - a vector of character strings with the characters present in the font
\item \code{png} - a matrix of 0's and 1's grabbed from the PNG file for the font using the package
\href{https://cran.r-project.org/package=imager}{imager}.
}
}
\examples{
data(font_sets)
}
\keyword{datasets}
