% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerCompassDir.R
\name{getPlayerCompassDir}
\alias{getPlayerCompassDir}
\title{Determine player's compass direction}
\usage{
getPlayerCompassDir(player_id = NULL, method = c("rotation", "direction"))
}
\arguments{
\item{player_id}{Player ID}

\item{method}{get player direction either from
\code{\link[miner:getPlayerRotation]{miner::getPlayerRotation()}} or \code{\link[miner:getPlayerDirection]{miner::getPlayerDirection()}}.}
}
\value{
Character string indication player direction
}
\description{
Figure out the largest value of a player's direction and translate
to a compass direction.
}
\author{
Felix Ling
}
