% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlocksMix.R
\name{setBlocksMix}
\alias{setBlocksMix}
\title{Place a random mixture of blocks in a cuboid}
\usage{
setBlocksMix(x0, y0, z0, x1, y1, z1, ids, styles = NULL, prob = NULL)
}
\arguments{
\item{x0}{A numeric string with north/south position of one corner}

\item{y0}{A numeric string with height of one corner}

\item{z0}{A numeric string with east/west position of one corner}

\item{x1}{A numeric string with north/south position of opposite corner}

\item{y1}{A numeric string with height of opposite corner}

\item{z1}{A numeric string with east/west position of opposite corner}

\item{ids}{Vector of block ids}

\item{styles}{Vector of block styles (same length as \code{ids})}

\item{prob}{Probabilities for each block type (same length as \code{ids})
If NULL, they are taken as equally probable.}
}
\value{
None.
}
\description{
Place blocks of a random mixture of types in the cuboid
with opposite corners at the positions (x0, y0, z0) and (x1, y1, z1).
}
\details{
This is like \code{\link[miner:setBlocks]{miner::setBlocks()}} but placing a random
mixture of blocks.
}
\examples{
\dontrun{
mc_connect()

items <- find_item("stained glass$")
pos <- whereami()
setBlocksMix(pos[1]+2, pos[2], pos[3]+2,
             pos[1]+2, pos[2]+5, pos[3]+2,
             ids=items[,2], styles=items[,3])
}

}
\seealso{
\code{\link[miner:setBlock]{miner::setBlock()}}, \code{\link[miner:setBlocks]{miner::setBlocks()}}, \code{\link[=setBlocksStyle]{setBlocksStyle()}}
}
\author{
Felix Ling
}
