% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlocksStyle.R
\name{setBlocksStyle}
\alias{setBlocksStyle}
\title{Place blocks in a cuboid}
\usage{
setBlocksStyle(x0, y0, z0, x1, y1, z1, id, style = 0)
}
\arguments{
\item{x0}{A numeric string with north/south position of one corner}

\item{y0}{A numeric string with height of one corner}

\item{z0}{A numeric string with east/west position of one corner}

\item{x1}{A numeric string with north/south position of opposite corner}

\item{y1}{A numeric string with height of opposite corner}

\item{z1}{A numeric string with east/west position of opposite corner}

\item{id}{Block id}

\item{style}{Block style}
}
\value{
None.
}
\description{
Place blocks of a single type (specified by \code{id}) in the cuboid
with opposite corners at the positions (x0, y0, z0) and (x1, y1, z1).
}
\details{
This is just like \code{\link[miner:setBlocks]{miner::setBlocks()}} but with an
added \code{style} argument.
}
\examples{
\dontrun{
mc_connect()

item <- find_item("Blue Stained Glass")
pos <- whereami()
setBlocksStyle(pos[1]+2, pos[2], pos[3]+2,
          pos[1]+2, pos[2]+5, pos[3]+2,
          id=item[2], style=item[3])
}

}
\seealso{
\code{\link[miner:setBlock]{miner::setBlock()}}, \code{\link[miner:setBlocks]{miner::setBlocks()}}
}
