% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlockV.R
\name{setPlayerDirectionV}
\alias{setPlayerDirectionV}
\title{setPlayerDirection but taking a vector of positions}
\usage{
setPlayerDirectionV(pos, player_id = NULL)
}
\arguments{
\item{pos}{Vector (x,y,z) giving position}

\item{player_id}{Player ID}
}
\value{
None.
}
\description{
Rotate player to direction (x,y,z) specified as a vector
}
\examples{
\dontrun{
library(miner)
mc_connect()
setPlayerDirectionV( c(1, 0, 1) )
}
}
\seealso{
\code{\link[miner:setPlayerDirection]{miner::setPlayerDirection()}}
}
