% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereami.R
\name{whereami}
\alias{whereami}
\title{Get rounded player position}
\usage{
whereami(player_id = NULL, tile = FALSE)
}
\arguments{
\item{player_id}{Integer giving the ID of a player}

\item{tile}{If TRUE, truncate the result to integers}
}
\description{
Get rounded player position
}
\details{
This is just like \code{\link[miner:getPlayerPos]{miner::getPlayerPos()}} but with \code{tile=TRUE}
being the default.
}
\examples{
\dontrun{
mc_connect()
whereami()
}
}
\seealso{
\code{\link[miner:getPlayerPos]{miner::getPlayerPos()}}
}
\author{
Felix Ling
}
